requirements.txt
africastalking==1.1.9
astroid==2.3.1
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
contextlib2==0.5.5
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.2
python-dotenv==0.10.3
requests==2.22.0
schema==0.7.1
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
